+++ backend/src/main/java/groom/backend/interfaces/coupon/CouponController.java
@@ -1,5 +1,6 @@
 package groom.backend.interfaces.coupon;
 
+import groom.backend.application.coupon.CouponIssueService;
 import groom.backend.application.coupon.CouponService;
 import groom.backend.domain.auth.entity.User;
 import groom.backend.infrastructure.security.CustomUserDetails;
@@ -30,6 +31,8 @@ import java.util.List;
 @RequestMapping("/coupon")
 public class CouponController {
   private final CouponService couponService;
+  private final CouponIssueService couponIssueService;
+
   /**
    * 쿠폰 생성
    * POST /coupon
@@ -122,7 +125,7 @@ public class CouponController {
     }
 
     // 쿠폰 발급
-    CouponIssueResponse response = couponService.issueCoupon(couponId, user);
+    CouponIssueResponse response = couponIssueService.issueCoupon(couponId, user);
 
     // 쿠폰이 존재하지 않을 시
     if (response == null) {
@@ -145,7 +148,7 @@ public class CouponController {
     Long userId = user.getId();
 
     // 내 미사용 쿠폰 조회
-    List<CouponIssueResponse> response = couponService.searchMyCoupon(userId);
+    List<CouponIssueResponse> response = couponIssueService.searchMyCoupon(userId);
 
     return ResponseEntity.ok(response);
   }
